Consolidate spec helpers
