+++ spec/spec_helper.rb
@@ -1,71 +1,126 @@
-# This file was generated by the `rails generate rspec:install` command. Conventionally, all
-# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
-# The generated `.rspec` file contains `--require spec_helper` which will cause this
-# file to always be loaded, without a need to explicitly require it in any files.
-#
-# Given that it is always loaded, you are encouraged to keep this file as
-# light-weight as possible. Requiring heavyweight dependencies from this file
-# will add to the boot time of your test suite on EVERY test run, even for an
-# individual file that may not need all of that loaded. Instead, make a
-# separate helper file that requires this one and then use it only in the specs
-# that actually need it.
-#
-# The `.rspec` file also contains a few flags that are not defaults but that
-# users commonly want.
-#
-# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+if ENV["COVERAGE"]
+  require 'simplecov'
+  SimpleCov.start 'rails'
+end
+
+ENV['RAILS_ENV'] ||= 'test'
+require File.expand_path('../../config/environment', __FILE__)
+require 'rspec/rails'
+require 'factory_girl'
+require 'rexml/document'
+FactoryGirl.find_definitions
+
+class EmailNotify
+  class << self
+    alias real_send_user_create_notification send_user_create_notification
+    def send_user_create_notification _user; end
+  end
+end
+
+class ActionView::TestCase::TestController
+  include Rails.application.routes.url_helpers
+end
+
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+
+module RSpec
+  module Core
+    module Hooks
+      class HookCollection
+        def find_hooks_for(group)
+          self.class.new(select { |hook| hook.options_apply?(group) })
+        end
+      end
+    end
+  end
+end
+
 RSpec.configure do |config|
-  # These two settings work together to allow you to limit a spec run
-  # to individual examples or groups you care about by tagging them with
-  # `:focus` metadata. When nothing is tagged with `:focus`, all examples
-  # get run.
+  config.use_transactional_fixtures = true
+  config.use_instantiated_fixtures  = false
+  config.fixture_path = "#{::Rails.root}/test/fixtures"
+  config.infer_spec_type_from_file_location!
   config.filter_run :focus
   config.run_all_when_everything_filtered = true
+  config.order = :random
 
-  # Many RSpec users commonly either run the entire suite or an individual
-  # file, and it's useful to allow more verbose output when running an
-  # individual spec file.
+  Kernel.srand config.seed
+
+  config.include SessionHelpers, type: :feature
+
+  # shortcuts for factory_girl to use: create / build / build_stubbed
+  config.include FactoryGirl::Syntax::Methods
+
   if config.files_to_run.one?
-    # Use the documentation formatter for detailed output,
-    # unless a formatter has already been configured
-    # (e.g. via a command-line flag).
     config.default_formatter = 'doc'
   end
 
-  # Run specs in random order to surface order dependencies. If you find an
-  # order dependency and want to debug it, you can fix the order by providing
-  # the seed, which is printed after each run.
-  #     --seed 1234
-  config.order = :random
-
-  # Seed global randomization in this process using the `--seed` CLI option.
-  # Setting this allows you to use `--seed` to deterministically reproduce
-  # test failures related to randomization by passing the same `--seed` value
-  # as the one that triggered the failure.
-  Kernel.srand config.seed
-
-  # rspec-expectations config goes here. You can use an alternate
-  # assertion/expectation library such as wrong or the stdlib/minitest
-  # assertions if you prefer.
   config.expect_with :rspec do |expectations|
-    # Enable only the newer, non-monkey-patching expect syntax.
-    # For more details, see:
-    #   - http://myronmars.to/n/dev-blog/2012/06/rspecs-new-expectation-syntax
     expectations.syntax = :expect
   end
 
-  # rspec-mocks config goes here. You can use an alternate test double
-  # library (such as bogus or mocha) by changing the `mock_with` option here.
   config.mock_with :rspec do |mocks|
-    # Enable only the newer, non-monkey-patching expect syntax.
-    # For more details, see:
-    #   - http://teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
     # TODO: Disable :should syntax
     mocks.syntax = [:expect, :should]
-
-    # Prevents you from mocking or stubbing a method that does not exist on
-    # a real object. This is generally recommended.
     # TODO: Enable this and fix the resulting problems
     # mocks.verify_partial_doubles = true
   end
+end
+
+def define_spec_public_cache_directory
+  ActionController::Base.page_cache_directory = File.join(Rails.root, 'spec', 'public')
+  unless File.exist? ActionController::Base.page_cache_directory
+    FileUtils.mkdir_p ActionController::Base.page_cache_directory
+  end
+end
+
+def path_for_file_in_spec_public_cache_directory(file)
+  define_spec_public_cache_directory
+  File.join(ActionController::Base.page_cache_directory, file)
+end
+
+def create_file_in_spec_public_cache_directory(file)
+  file_path = path_for_file_in_spec_public_cache_directory(file)
+  File.open(file_path, 'a').close
+  file_path
+end
+
+# TODO: Clean up use of these Test::Unit style expectations
+def assert_xml(xml)
+  expect do
+    assert REXML::Document.new(xml)
+  end.not_to raise_error
+end
+
+def assert_atom10 feed, count
+  doc = Nokogiri::XML.parse(feed)
+  root = doc.css(':root').first
+  expect(root.name).to eq('feed')
+  expect(root.namespace.href).to eq('http://www.w3.org/2005/Atom')
+  expect(root.css('entry').count).to eq(count)
+end
+
+def assert_rss20 feed, count
+  doc = Nokogiri::XML.parse(feed)
+  root = doc.css(':root').first
+  expect(root.name).to eq('rss')
+  expect(root['version']).to eq('2.0')
+  expect(root.css('channel item').count).to eq(count)
+end
+
+def stub_full_article(time = Time.now)
+  author = FactoryGirl.build_stubbed(User, name: 'User Name')
+  text_filter = FactoryGirl.build(:textile)
+
+  a = FactoryGirl.build_stubbed(Article, published_at: time, user: author,
+                 created_at: time, updated_at: time,
+                 title: 'Foo Bar', permalink: 'foo-bar',
+                 guid: time.hash)
+  allow(a).to receive(:published_comments) { [] }
+  allow(a).to receive(:resources) { [FactoryGirl.build(:resource)] }
+  allow(a).to receive(:tags) { [FactoryGirl.build(:tag)] }
+  allow(a).to receive(:text_filter) { text_filter }
+  a
 end
